# document/spec/openapi.yaml
openapi: 3.1.0
info:
  title: lib4gz LMS Platform Backend API
  description: >
    RESTful API for the lib4gz Learning Management System platform.
    Provides endpoints for user authentication, course management, enrollments,
    modules, lessons, summaries, exercises, questions, and submissions.
  version: 1.0.0
  contact:
    name: lib4gz Team
servers:
  - url: http://localhost:8080/api
    description: Local Development
security:
  - ApiKeyAuth: []
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: access_token
      description: JWT access token
  parameters:
    XUserIdHeader:
      name: X-User-Id
      in: header
      required: true
      description: TypeID of the authenticated user, injected by JwtAuthFilter
      schema:
        type: string
        format: typeid
...
    ExerciseIdPath:
      name: exerciseId
      in: path
      required: true
      description: TypeID of the exercise
      schema:
        type: string
        format: typeid
...
    SubmissionIdPath:
      name: submissionId
      in: path
      required: true
      description: TypeID of the submission
      schema:
        type: string
        format: typeid
    AnswerIdPath:
      name: answerId
      in: path
      required: true
      description: TypeID of the answer
      schema:
        type: string
        format: typeid
  schemas:
...
    QuestionVisibility:
      type: string
      enum:
        - VISIBLE
        - HIDDEN
    QuestionAction:
      type: string
      enum:
        - CREATE
        - UPDATE
        - DELETE
    SubmissionStatus:
      type: string
      enum:
        - DRAFT
        - SUBMITTED
        - APPROVED
        - NEEDS_REVISION
...
    UserSummary:
      type: object
      properties:
        id:
          type: string
          format: typeid
        name:
          type: string
        email:
          type: string
        avatarUrl:
          type:
            - string
            - "null"

    # ---------- Courses ----------
...
    CreateQuestionsRequest:
      type: object
      required:
        - questions
      properties:
        questions:
          type: array
          items:
            $ref: "#/components/schemas/CreateQuestionItem"
    CreateQuestionItem:
      type: object
      required:
        - content
      properties:
        content:
          type: string
        orderIndex:
          type:
            - integer
            - "null"
        meta:
          type: object
          default: {}
        visibility:
          $ref: "#/components/schemas/QuestionVisibility"
          default: VISIBLE
    UpdateQuestionsRequest:
      type: object
      required:
        - questions
      properties:
        questions:
          type: array
          items:
            $ref: "#/components/schemas/UpdateQuestionItem"
    UpdateQuestionItem:
      type: object
      required:
        - action
      properties:
        action:
          $ref: "#/components/schemas/QuestionAction"
        questionId:
          type:
            - string
            - "null"
          format: typeid
        content:
          type:
            - string
            - "null"
        orderIndex:
          type:
            - integer
            - "null"
        meta:
          type:
            - object
            - "null"
        visibility:
          oneOf:
            - $ref: "#/components/schemas/QuestionVisibility"
            - type: "null"
    QuestionResponse:
      type: object
      properties:
        id:
          type: string
          format: typeid
        exerciseId:
          type: string
          format: typeid
        content:
          type: string
        orderIndex:
          type: integer
        meta:
          type: object
        visibility:
          $ref: "#/components/schemas/QuestionVisibility"
        createdAt:
          type: integer
          format: int64
        updatedAt:
          type: integer
          format: int64
    QuestionsResponse:
      type: object
      properties:
        created:
          type: array
          items:
            $ref: "#/components/schemas/QuestionResponse"
        updated:
          type: array
          items:
            $ref: "#/components/schemas/QuestionResponse"
        deleted:
          type: array
          items:
            type: string
            format: typeid

    # ---------- Submissions ----------
    CreateSubmissionRequest:
      type: object
      properties:
        answers:
          type: array
          items:
            $ref: "#/components/schemas/AnswerRequest"
          default: []
...
    AnswerRequest:
      type: object
      required:
        - questionId
      properties:
        questionId:
          type: string
          format: typeid
        answer:
          type:
            - string
            - "null"
    SubmissionResponse:
      type: object
      properties:
        id:
          type: string
          format: typeid
        exerciseId:
          type: string
          format: typeid
        user:
          $ref: "#/components/schemas/UserSummary"
        status:
          $ref: "#/components/schemas/SubmissionStatus"
        answers:
          type:
            - array
            - "null"
          items:
            $ref: "#/components/schemas/StudentAnswerResponse"
        createdAt:
          type: integer
          format: int64
        updatedAt:
          type: integer
          format: int64
        submittedAt:
          type:
            - integer
            - "null"
          format: int64
        approvedAt:
          type:
            - integer
            - "null"
          format: int64
    StudentAnswerResponse:
      type: object
      properties:
        id:
          type: string
          format: typeid
        questionId:
          type: string
          format: typeid
        questionContent:
          type: string
        answer:
          type:
            - string
            - "null"
        teacherComment:
          type:
            - string
            - "null"
        createdAt:
          type: integer
          format: int64
        updatedAt:
          type: integer
          format: int64
    RevisionRequest:
      type: object
      properties:
        feedback:
          type:
            - string
            - "null"
    CommentRequest:
      type: object
      required:
        - comment
      properties:
        comment:
          type: string

    # ---------- Errors ----------
    ErrorResp:
      type: object
      properties:
        timestamp:
          type: integer
          format: int64
        path:
          type: string
        status:
          type: integer
        requestId:
          type: string
        traceId:
          type: string
        errors:
          type: array
          items:
            $ref: "#/components/schemas/ApiError"
    ApiError:
      oneOf:
        - $ref: "#/components/schemas/InputError"
        - $ref: "#/components/schemas/GenericError"
      discriminator:
        propertyName: type
    InputError:
      type: object
      properties:
        type:
          type: string
          enum:
            - InputError
        field:
          type: string
        message:
          type: string
    GenericError:
      type: object
      properties:
        type:
          type: string
          enum:
            - SystemError
        reason:
          type: string
  responses:
    BadRequest:
      description: Bad request - validation errors or malformed input
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResp"
    Unauthorized:
      description: Unauthorized - missing or invalid access token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResp"
    Forbidden:
      description: Forbidden - insufficient permissions
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResp"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResp"
paths:
...
  /v1/exercises/{exerciseId}/questions:
    get:
      tags:
        - questions
      operationId: listQuestions
      summary: List questions for an exercise
      description: Retrieves all questions for the specified exercise, ordered by orderIndex.
      parameters:
        - $ref: "#/components/parameters/XUserIdHeader"
        - $ref: "#/components/parameters/ExerciseIdPath"
      responses:
        "200":
          description: List of questions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/QuestionResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"

    post:
      tags:
        - questions
      operationId: createQuestions
      summary: Create questions for an exercise
      description: Creates one or more questions for the specified exercise in a single batch. Restricted to teachers and course creators.
      parameters:
        - $ref: "#/components/parameters/XUserIdHeader"
        - $ref: "#/components/parameters/ExerciseIdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateQuestionsRequest"
      responses:
        "201":
          description: Questions created successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/QuestionResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"

    patch:
      tags:
        - questions
      operationId: updateQuestions
      summary: Batch update questions for an exercise
      description: >
        Performs batch operations (create, update, delete) on questions for the specified exercise.
        Each item in the request specifies an action to perform. Restricted to teachers and course creators.
      parameters:
        - $ref: "#/components/parameters/XUserIdHeader"
        - $ref: "#/components/parameters/ExerciseIdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateQuestionsRequest"
      responses:
        "200":
          description: Questions updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuestionsResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"

  # ========================================================================
  # SUBMISSIONS
  # ========================================================================
  /v1/exercises/{exerciseId}/submissions:
    get:
      tags:
        - submissions
      operationId: listSubmissions
      summary: List submissions for an exercise
      description: >
        Retrieves all submissions for the specified exercise.
        Teachers see all submissions; learners may only see their own depending on settings.
      parameters:
        - $ref: "#/components/parameters/XUserIdHeader"
        - $ref: "#/components/parameters/ExerciseIdPath"
      responses:
        "200":
          description: List of submissions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SubmissionResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"

    post:
      tags:
        - submissions
      operationId: createSubmission
      summary: Create a new submission
      description: >
        Creates a new draft submission for the specified exercise.
        The authenticated user becomes the submission owner.
      parameters:
        - $ref: "#/components/parameters/XUserIdHeader"
        - $ref: "#/components/parameters/ExerciseIdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSubmissionRequest"
      responses:
        "201":
          description: Submission created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubmissionResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /v1/exercises/{exerciseId}/my-submission:
    get:
      tags:
        - submissions
      operationId: getMySubmission
      summary: Get current user submission for an exercise
      description: Retrieves the submission record for the authenticated user for the specified exercise.
      parameters:
        - $ref: "#/components/parameters/XUserIdHeader"
        - $ref: "#/components/parameters/ExerciseIdPath"
      responses:
        "200":
          description: Submission details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubmissionResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /v1/submissions/{submissionId}:
    get:
      tags:
        - submissions
      operationId: getSubmission
      summary: Get a submission by ID
      description: Retrieves the details of a specific submission including all answers.
      parameters:
        - $ref: "#/components/parameters/XUserIdHeader"
        - $ref: "#/components/parameters/SubmissionIdPath"
      responses:
        "200":
          description: Submission details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubmissionResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /v1/submissions/{submissionId}/submit:
    post:
      tags:
        - submissions
      operationId: submitSubmission
      summary: Submit a draft submission for review
      description: >
        Transitions a submission from DRAFT to SUBMITTED status.
        Once submitted, the learner cannot edit their answers until revision is requested.
      parameters:
        - $ref: "#/components/parameters/XUserIdHeader"
        - $ref: "#/components/parameters/SubmissionIdPath"
      responses:
        "200":
          description: Submission submitted successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubmissionResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /v1/submissions/{submissionId}/approve:
    post:
      tags:
        - submissions
      operationId: approveSubmission
      summary: Approve a submitted submission
      description: >
        Transitions a submission from SUBMITTED to APPROVED status.
        Restricted to teachers and course creators.
      parameters:
        - $ref: "#/components/parameters/XUserIdHeader"
        - $ref: "#/components/parameters/SubmissionIdPath"
      responses:
        "200":
          description: Submission approved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubmissionResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /v1/submissions/{submissionId}/revision:
    post:
      tags:
        - submissions
      operationId: requestRevision
      summary: Request revision of a submission
      description: >
        Transitions a submission from SUBMITTED to NEEDS_REVISION status.
        Optionally includes feedback for the learner. Restricted to teachers and course creators.
      parameters:
        - $ref: "#/components/parameters/XUserIdHeader"
        - $ref: "#/components/parameters/SubmissionIdPath"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RevisionRequest"
      responses:
        "200":
          description: Revision requested successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubmissionResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /v1/answers/{answerId}/comment:
    post:
      tags:
        - submissions
      operationId: commentOnAnswer
      summary: Add a teacher comment to an answer
      description: >
        Adds or updates a teacher comment on a specific student answer.
        Restricted to teachers and course creators.
      parameters:
        - $ref: "#/components/parameters/XUserIdHeader"
        - $ref: "#/components/parameters/AnswerIdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CommentRequest"
      responses:
        "200":
          description: Comment added successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubmissionResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"